Extract motion, path, goal, and purpose chains
